Project and Layout
Created a project for the code and updated with the required layouts.

On the LocationFeedListView, I ended up using a border container
because the list containers did not render the same as the wireframe.
When using the exact code below from the Adobe Flex 4.6 site in
IntelliJ, the layout looked different from the website rendering and
wireframe.

<s:List>
        <mx:ArrayCollection>
            <fx:String>Flash</fx:String>
            <fx:String>Director</fx:String>
            <fx:String>Dreamweaver</fx:String>
            <fx:String>ColdFusion</fx:String>
        </mx:ArrayCollection>
    </s:List>